ML/requirements.txt
## The following requirements were added by pip freeze:
click==8.0.1
cycler==0.10.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.1
opencv-python-headless==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
scipy==1.7.1
six==1.16.0
Werkzeug==2.0.1
